:lipstick:Adiciona diálogo de remoção de tarefa

+++ lib/view/tarefaEditPage.dart
@@ -42,9 +42,45 @@ class _NoteEditPageState extends State<TarefaEditPage> implements INoteEdit {
     this.presenter.updateNote(note);
   }
 
+  void _showDialogDelete() {
+    showDialog(
+        context: context,
+        //barrierDismissible: false,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: Text(
+              "Remoção",
+            ),
+            content: Text(
+              "Confirmar exclusão?",
+            ),
+            actions: <Widget>[
+              OutlineButton(
+                  child: Text(
+                    "Cancelar",
+                    style: TextStyle(fontSize: 15),
+                  ),
+                  onPressed: () {
+                    Navigator.of(context).pop();
+                  }),
+              RaisedButton(
+                  color: Theme.of(context).primaryColor,
+                  child: Text(
+                    "Excluir",
+                    style: TextStyle(fontSize: 15),
+                  ),
+                  onPressed: () {
+                    this.presenter.delete(widget.note["id"]);
+                    Navigator.of(context).pop();
+                  })
+            ],
+          );
+        });
+  }
+
   @override
   void onClickDelete() {
-    this.presenter.delete(widget.note["id"]);
+    this._showDialogDelete();
   }
 
   @override
@@ -69,191 +105,191 @@ class _NoteEditPageState extends State<TarefaEditPage> implements INoteEdit {
 
   @override
   Widget build(BuildContext context) {
-    return   
-     Scaffold(
-          appBar: AppBar(
-            title: Text("Editar tarefa"),
-            actions: <Widget>[
-              IconButton(
-                  icon: Icon(Icons.save, color: Colors.white),
-                  onPressed: () {
-                    if (_formKey.currentState.validate()) {
-                      this.onClickUpdate();
-                    }
-                  }),
-              IconButton(
-                  icon: Icon(Icons.delete),
-                  onPressed: () {
-                    this.onClickDelete();
-                  }),
-            ],
-            leading: Builder(builder: (BuildContext context) {
-              return IconButton(
-                icon: Icon(Icons.arrow_back, color: Colors.white),
-                onPressed: () {
-                  this.backPage();
-                },
-              );
-            }),
-          ),
-          body:
-            SafeArea(
-            child: Builder(builder: (context) {
-            this.scaffoldContext = context;
-            return Form(
-              key: _formKey,
-              child: SingleChildScrollView(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Editar tarefa"),
+        actions: <Widget>[
+          IconButton(
+              icon: Icon(Icons.save, color: Colors.white),
+              onPressed: () {
+                if (_formKey.currentState.validate()) {
+                  this.onClickUpdate();
+                }
+              }),
+          IconButton(
+              icon: Icon(Icons.delete),
+              onPressed: () {
+                this.onClickDelete();
+              }),
+        ],
+        leading: Builder(builder: (BuildContext context) {
+          return IconButton(
+            icon: Icon(Icons.arrow_back, color: Colors.white),
+            onPressed: () {
+              this.backPage();
+            },
+          );
+        }),
+      ),
+      body: SafeArea(child: Builder(builder: (context) {
+        this.scaffoldContext = context;
+        return Form(
+          key: _formKey,
+          child: SingleChildScrollView(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: <Widget>[
+                Container(
+                  padding: EdgeInsets.fromLTRB(15, 30, 15, 5),
+                  child: TextFormField(
+                    validator: (value) {
+                      if (value.isEmpty) {
+                        return "Preencha o campo titulo";
+                      }
+                      return null;
+                    },
+                    controller: this._txtTitle,
+                    decoration: InputDecoration(
+                        labelText: "Título", border: OutlineInputBorder()),
+                  ),
+                ),
+                Container(
+                  padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
+                  child: TextFormField(
+                    maxLines: 10,
+                    controller: this._txtDescription,
+                    decoration: InputDecoration(
+                        labelText: "Descrição", border: OutlineInputBorder()),
+                    validator: (value) {
+                      if (value.isEmpty) {
+                        return "Preencha o campo descrição";
+                      }
+                      return null;
+                    },
+                  ),
+                ),
+                Row(
                   children: <Widget>[
-                    Container(
-                      padding: EdgeInsets.fromLTRB(15, 30, 15, 5),
-                      child: TextFormField(
-                        validator: (value) {
-                          if (value.isEmpty) {
-                            return "Preencha o campo titulo";
-                          }
-                          return null;
-                        },
-                        controller: this._txtTitle,
-                        decoration: InputDecoration(
-                            labelText: "Título", border: OutlineInputBorder()),
-                      ),
-                    ),
-                    Container(
-                      padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
-                      child: TextFormField(
-                        maxLines: 10,
-                        controller: this._txtDescription,
-                        decoration: InputDecoration(
-                            labelText: "Descrição",
-                            border: OutlineInputBorder()),
-                        validator: (value) {
-                          if (value.isEmpty) {
-                            return "Preencha o campo descrição";
-                          }
-                          return null;
-                        },
-                      ),
-                    ),
-                    Row(
-                      children: <Widget>[
-                        Expanded(
-                          flex: 8,
-                          child: Container(
-                              padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
-                              child: TextFormField(
-                                controller: this._txtDataStart,
-                                decoration: InputDecoration(
-                                    hintText: "00/00/0000",
-                                    labelText: "Data de ínicio",
-                                    border: OutlineInputBorder()),
-                                validator: (value) {
-                                  if (value.isEmpty) {
-                                    return "Preencha o campo data de ínicio";
-                                  }
-                                  if(!Validation.isDateValida(value)){
-                                    return "Data inválida";
-                                  }
-                                  return null;
-                                },
-                              )),
-                        ),
-                        Expanded(
-                            flex: 1,
-                            child: IconButton(
-                              icon: Icon(Icons.calendar_today,
-                                  color: Theme.of(context).primaryColor),
-                              onPressed: () {
-                                setState(() {
-                                  this._selectDateStart(context);
-                                });
-                              },
-                            )),
-                      ],
-                    ),
-                    Row(
-                      children: <Widget>[
-                        Expanded(
-                          flex: 8,
-                          child: Container(
-                              padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
-                              child: TextFormField(
-                                controller: this._txtDateEnd,
-                                decoration: InputDecoration(
-                                    hintText: "00/00/0000",
-                                    labelText: "Data de término",
-                                    border: OutlineInputBorder()),
-                                validator: (value) {
-                                  if (value.isEmpty) {
-                                    return "Preencha o campo data de término";
-                                  }
-                                  if(!Validation.isDateValida(value)){
-                                    return "Data inválida";
-                                  }
-                                  if(Validation.isDateValida(this._txtDataStart.text)){
-                                    DateTime dateStart = DateConversion.dateFormtToDateTime(this._txtDataStart.text);
-                                    DateTime dateEnd = DateConversion.dateFormtToDateTime(this._txtDateEnd.text);
-                                    if(dateEnd.compareTo(dateStart) < 0){
-                                      return "Data anterior a data de ínicio";
-                                    }
-                                  }
-                                  return null;
-                                },
-                              )),
-                        ),
-                        Expanded(
-                            flex: 1,
-                            child: IconButton(
-                              icon: Icon(
-                                Icons.calendar_today,
-                                color: Theme.of(context).primaryColor,
-                              ),
-                              onPressed: () {
-                                setState(() {
-                                  this._selectDateEnd(context);
-                                });
-                              },
-                            )),
-                      ],
-                    ),
-                    Container(
-                      padding: EdgeInsets.fromLTRB(15, 10, 15, 10),
-                      child: DropdownButton(                     
-                          isExpanded: true,
-                          value: this.prioridadeSelected,
-                          items: this.prioridades.map((dropDownValue) {
-                            return DropdownMenuItem<String>(
-                                value: dropDownValue,
-                                child: Text(dropDownValue));
-                          }).toList(),
-                          onChanged: (newValue) {
-                            this.onDropDownItemSelected(newValue);
-                          }),
-                    ),
-                    Row(
-                      children: <Widget>[
-                        Container(
+                    Expanded(
+                      flex: 8,
+                      child: Container(
                           padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
-                          child: Text(
-                            "Concluido",
-                            style: TextStyle(color: Colors.grey),
-                          ),
-                        ),
-                        Checkbox(
-                            value: this._checkDone,
-                            onChanged: (newValue) {
-                              setState(() {
-                                this._checkDone = newValue;
-                              });
-                            }),
-                      ],
+                          child: TextFormField(
+                            controller: this._txtDataStart,
+                            decoration: InputDecoration(
+                                hintText: "00/00/0000",
+                                labelText: "Data de ínicio",
+                                border: OutlineInputBorder()),
+                            validator: (value) {
+                              if (value.isEmpty) {
+                                return "Preencha o campo data de ínicio";
+                              }
+                              if (!Validation.isDateValida(value)) {
+                                return "Data inválida";
+                              }
+                              return null;
+                            },
+                          )),
                     ),
+                    Expanded(
+                        flex: 1,
+                        child: IconButton(
+                          icon: Icon(Icons.calendar_today,
+                              color: Theme.of(context).primaryColor),
+                          onPressed: () {
+                            setState(() {
+                              this._selectDateStart(context);
+                            });
+                          },
+                        )),
                   ],
                 ),
-              ),
-            );
-          })),
+                Row(
+                  children: <Widget>[
+                    Expanded(
+                      flex: 8,
+                      child: Container(
+                          padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
+                          child: TextFormField(
+                            controller: this._txtDateEnd,
+                            decoration: InputDecoration(
+                                hintText: "00/00/0000",
+                                labelText: "Data de término",
+                                border: OutlineInputBorder()),
+                            validator: (value) {
+                              if (value.isEmpty) {
+                                return "Preencha o campo data de término";
+                              }
+                              if (!Validation.isDateValida(value)) {
+                                return "Data inválida";
+                              }
+                              if (Validation.isDateValida(
+                                  this._txtDataStart.text)) {
+                                DateTime dateStart =
+                                    DateConversion.dateFormtToDateTime(
+                                        this._txtDataStart.text);
+                                DateTime dateEnd =
+                                    DateConversion.dateFormtToDateTime(
+                                        this._txtDateEnd.text);
+                                if (dateEnd.compareTo(dateStart) < 0) {
+                                  return "Data anterior a data de ínicio";
+                                }
+                              }
+                              return null;
+                            },
+                          )),
+                    ),
+                    Expanded(
+                        flex: 1,
+                        child: IconButton(
+                          icon: Icon(
+                            Icons.calendar_today,
+                            color: Theme.of(context).primaryColor,
+                          ),
+                          onPressed: () {
+                            setState(() {
+                              this._selectDateEnd(context);
+                            });
+                          },
+                        )),
+                  ],
+                ),
+                Container(
+                  padding: EdgeInsets.fromLTRB(15, 10, 15, 10),
+                  child: DropdownButton(
+                      isExpanded: true,
+                      value: this.prioridadeSelected,
+                      items: this.prioridades.map((dropDownValue) {
+                        return DropdownMenuItem<String>(
+                            value: dropDownValue, child: Text(dropDownValue));
+                      }).toList(),
+                      onChanged: (newValue) {
+                        this.onDropDownItemSelected(newValue);
+                      }),
+                ),
+                Row(
+                  children: <Widget>[
+                    Container(
+                      padding: EdgeInsets.fromLTRB(15, 10, 15, 5),
+                      child: Text(
+                        "Concluido",
+                        style: TextStyle(color: Colors.grey),
+                      ),
+                    ),
+                    Checkbox(
+                        value: this._checkDone,
+                        onChanged: (newValue) {
+                          setState(() {
+                            this._checkDone = newValue;
+                          });
+                        }),
+                  ],
+                ),
+              ],
+            ),
+          ),
+        );
+      })),
     );
   }
